+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Removed
+- Default Plugins ([#70](https://github.com/cloudogu/scm/pull/70/files))
+
 ## [2.42.1-1]
 ### Changed
 - Upgrade SCM-Manager to version 2.42.1 ([Changelog](https://github.com/scm-manager/scm-manager/blob/2.42.1/CHANGELOG.md))